+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Implementation changes and bug fixes
 - [PR #1638](https://github.com/rqlite/rqlite/pull/1638): More Snapshotting metrics.
 - [PR #1639](https://github.com/rqlite/rqlite/pull/1639): Reuse Session and S3 manager when working with AWS S3. Thanks @mauri870
+- [PR #1640](https://github.com/rqlite/rqlite/pull/1640): Remove Restart checkpoint, tests show it's not an optimization.
 
 ## 8.18.2 (January 27th 2024)
 ### Implementation changes and bug fixes